move generateStatus to main for use later
